Rename 'IO' namespaces to 'Internal'
