+++ SAGESharp/SAGESharp.csproj
@@ -56,15 +56,15 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SLB\ISLBBinaryReader.cs" />
     <Compile Include="SLB\Identifier.cs" />
-    <Compile Include="SLB\IO\IdentifierBinaryReader.cs" />
+    <Compile Include="SLB\Internal\IdentifierBinaryReader.cs" />
     <Compile Include="SLB\Level\Conversation\Character.cs" />
     <Compile Include="SLB\Level\Conversation\LineSide.cs" />
     <Compile Include="SLB\Level\Conversation\Frame.cs" />
     <Compile Include="SLB\Level\Conversation\Info.cs" />
-    <Compile Include="SLB\Level\Conversation\IO\CharacterBinaryReader.cs" />
-    <Compile Include="SLB\Level\Conversation\IO\ConversationBinaryReader.cs" />
-    <Compile Include="SLB\Level\Conversation\IO\FrameBinaryReader.cs" />
-    <Compile Include="SLB\Level\Conversation\IO\InfoBinaryReader.cs" />
+    <Compile Include="SLB\Level\Conversation\Internal\CharacterBinaryReader.cs" />
+    <Compile Include="SLB\Level\Conversation\Internal\ConversationBinaryReader.cs" />
+    <Compile Include="SLB\Level\Conversation\Internal\FrameBinaryReader.cs" />
+    <Compile Include="SLB\Level\Conversation\Internal\InfoBinaryReader.cs" />
     <Compile Include="SLB\SLBBinaryReaderFactory.cs" />
   </ItemGroup>
   <ItemGroup>
