+++ SAGESharp/SLB/Level/Conversation/Internal/CharacterBinaryReader.cs
@@ -2,7 +2,7 @@
 using System;
 using System.IO;
 
-namespace SAGESharp.SLB.Level.Conversation.IO
+namespace SAGESharp.SLB.Level.Conversation.Internal
 {
     /// <summary>
     /// Class to read Character objects from binary SLB files.
