+++ SAGESharpTests/SLB/Internal/IdentifierBinaryReaderTests.cs
@@ -1,12 +1,12 @@
 ﻿using Moq;
 using NUnit.Framework;
 using SAGESharp.SLB;
-using SAGESharp.SLB.IO;
+using SAGESharp.SLB.Internal;
 using SAGESharpTests.Util;
 using System;
 using System.IO;
 
-namespace SAGESharpTests.SLB.IO
+namespace SAGESharpTests.SLB.Internal
 {
     [TestFixture]
     public class IdentifierBinaryReaderTests
