+++ SAGESharpTests/SLB/Level/Conversation/Internal/CharacterBinaryReaderTests.cs
@@ -1,14 +1,13 @@
 ﻿using Moq;
 using NUnit.Framework;
 using SAGESharp.SLB;
-using SAGESharp.SLB.IO;
 using SAGESharp.SLB.Level.Conversation;
-using SAGESharp.SLB.Level.Conversation.IO;
+using SAGESharp.SLB.Level.Conversation.Internal;
 using SAGESharpTests.Util;
 using System;
 using System.IO;
 
-namespace SAGESharpTests.SLB.Level.Conversation.IO
+namespace SAGESharpTests.SLB.Level.Conversation.Internal
 {
     [TestFixture]
     public static class CharacterBinaryReaderTests
